LowkeyPlaces/backEnd/LowkeyPlaces/webApp/requirements.txt
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-generator==1.10
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.0.5
cryptography==37.0.2
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.8
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-filter==23.5
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.2
fonttools==4.37.4
fqdn==1.5.1
freezegun==1.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.13.0
idna==3.3
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.18.1
isoduration==20.11.0
isoweek==1.3.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
notebook_shim==0.2.3
numpy==1.23.3
outcome==1.1.0
overrides==7.4.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
Pillow==9.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.1.2
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
screen==1.0.1
selenium==4.1.5
Send2Trash==1.8.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.0
TimeConvert==3.0.13
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
trio==0.20.0
trio-websocket==0.9.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.9
virtualenv==20.25.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.1.0
xlwt==1.3.0
zipp==3.17.0
